Move BencodedInfo mapping logic into factory class
