+++ src/main/java/jtorrent/data/model/BencodedTorrent.java
@@ -24,6 +24,7 @@ import com.dampcake.bencode.BencodeInputStream;
 import jtorrent.data.model.exception.MappingException;
 import jtorrent.data.model.info.BencodedFile;
 import jtorrent.data.model.info.BencodedInfo;
+import jtorrent.data.model.info.BencodedInfoFactory;
 import jtorrent.domain.model.torrent.File;
 import jtorrent.domain.model.torrent.Sha1Hash;
 import jtorrent.domain.model.torrent.Torrent;
@@ -75,7 +76,7 @@ public class BencodedTorrent extends BencodedObject {
         String createdBy = getValueAsString(map, KEY_CREATED_BY).orElse("");
 
         Map<String, Object> infoDict = getValueAsMap(map, KEY_INFO);
-        BencodedInfo info = BencodedInfo.fromMap(infoDict);
+        BencodedInfo info = BencodedInfoFactory.fromMap(infoDict);
 
         return new BencodedTorrent(creationDate, announce, announceList, comment, createdBy, info);
     }
